refactor: move calendar functions to app level
